SoraSoundEffectManager的playDirect不再会cachefile了 SoraConsole添加setUseSysTerm以控制是否转发不能识别的命令到系统命令行 SoraAutoProfile现在使用SoraTimeStamp来计时 SoraCore的getCurrentSystemTime现在返回SoraTimestamp的currentTime, 以统一各个平台的单位 Serializable的serialize和unserialize接口改为接受更为强大的SoraStream而不是SoraMemoryBuffer了, 同时SoraNamedObject和SoraMemoryFile的相关函数也得到了改变 添加SoraFileStream，对于win32和posix的文件接口封装 SoraMemoryBuffer添加operator >>, <<和string <<, >> SoraTimestamp添加静态函数currentTime， 以方便使用

+++ Sora/Defaults/SoraTimer_OSX.h
@@ -65,8 +65,8 @@ namespace sora {
 				if(oldtime != time) {
 					oldtime = time;
 					time = mach_absolute_time();
-					fdt = dtToMil(time-oldtime) / 1000.0 * fTimeScale;
-					fps = 1.f / (dtToMil(time-oldtime)/1000.f);;
+					fdt = dtToMil(time-oldtime) / (float32)1000 * fTimeScale;
+					fps = 1.f / (dtToMil(time-oldtime)/(float32)1000);;
 				} else {
 					time = mach_absolute_time();
 				}
@@ -89,10 +89,10 @@ namespace sora {
 			time = mach_absolute_time();
 			
 			if(oldtime <= time - fpsInterval && oldtime >= time - (fpsInterval*2)) {
-				fdt = dtToMil(fpsInterval) / 1000.0 * fTimeScale;
+				fdt = dtToMil(fpsInterval) / (float32)1000 * fTimeScale;
 				oldtime += fpsInterval;
 			} else {
-				fdt = dtToMil(time-oldtime) / 1000.0 * fTimeScale;
+				fdt = dtToMil(time-oldtime) / (float32)1000 * fTimeScale;
 				oldtime = time;
 			}
 			if(fdt >= 1.f)
@@ -116,7 +116,7 @@ namespace sora {
 		}
 		
 		inline double dtToMil(uint64_t dt) {
-			return ((double)machTimeToNano(dt) / (1000.0*1000.0));
+			return ((double)machTimeToNano(dt) / (1000*1000));
 		}
 		
 		uint64_t time, oldtime;
